move logic to shared util
